pyproject.toml
[project]
name = "vlm-attention-viz"
version = "0.1.0"
description = "Visualize attention patterns in vision-language models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.51.0",
]

[project.optional-dependencies]
qwen = ["qwen-vl-utils>=0.0.8"]
dev = ["bertviz>=1.4.1", "ipykernel>=6.30.1"]

[project.scripts]
vlm-attention-viz = "vlm_attention_viz.cli:main"
